+++ lib/believe/models.rb
@@ -105,6 +105,8 @@ module Believe
 
   Match = Believe::Models::Match
 
+  MatchCompletedWebhookEvent = Believe::Models::MatchCompletedWebhookEvent
+
   MatchCreateParams = Believe::Models::MatchCreateParams
 
   MatchDeleteParams = Believe::Models::MatchDeleteParams
@@ -199,6 +201,8 @@ module Believe
 
   TeamMemberRetrieveParams = Believe::Models::TeamMemberRetrieveParams
 
+  TeamMemberTransferredWebhookEvent = Believe::Models::TeamMemberTransferredWebhookEvent
+
   TeamMemberUpdateParams = Believe::Models::TeamMemberUpdateParams
 
   TeamRetrieveParams = Believe::Models::TeamRetrieveParams
@@ -211,6 +215,8 @@ module Believe
 
   TurningPoint = Believe::Models::TurningPoint
 
+  UnwrapWebhookEvent = Believe::Models::UnwrapWebhookEvent
+
   VersionRetrieveParams = Believe::Models::VersionRetrieveParams
 
   WebhookCreateParams = Believe::Models::WebhookCreateParams
@@ -222,4 +228,6 @@ module Believe
   WebhookRetrieveParams = Believe::Models::WebhookRetrieveParams
 
   WebhookTriggerEventParams = Believe::Models::WebhookTriggerEventParams
+
+  WebhookUnwrapParams = Believe::Models::WebhookUnwrapParams
 end