+++ test/believe/resources/stream_test.rb
@@ -4,7 +4,7 @@ require_relative "../test_helper"
 
 class Believe::Test::Resources::StreamTest < Believe::Test::ResourceTest
   def test_test_connection
-    skip("Prism tests are disabled")
+    skip("Mock server tests are disabled")
 
     response = @believe.stream.test_connection