+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 0.1.0 (2026-02-20)
+
+Full Changelog: [v0.0.2...v0.1.0](https://github.com/cjavdev/believe-ruby/compare/v0.0.2...v0.1.0)
+
+### Features
+
+* **api:** Add webhook unwrapping ([2dd6323](https://github.com/cjavdev/believe-ruby/commit/2dd63230ca9016b848775a9f9453cb560110835a))
+* **api:** manual updates ([d96df4d](https://github.com/cjavdev/believe-ruby/commit/d96df4d175d5380700f66b85ed79751689755c79))
+
+
+### Chores
+
+* **internal:** remove mock server code ([6d33ca5](https://github.com/cjavdev/believe-ruby/commit/6d33ca5b9a2067e99fab6801c25f5e58d1dc7b43))
+* **test:** update skip reason message ([1af9158](https://github.com/cjavdev/believe-ruby/commit/1af9158034f6a76789a800b007486ec676c60f82))
+* update mock server docs ([393cc25](https://github.com/cjavdev/believe-ruby/commit/393cc2502c0f92208a55951aa38a33ddfbcea1ad))
+
 ## 0.0.2 (2026-02-11)
 
 Full Changelog: [v0.0.1...v0.0.2](https://github.com/cjavdev/believe-ruby/compare/v0.0.1...v0.0.2)